test(research): extract caption lane fixtures
